Add Person model
People can only have one set of ContactDetails, but ContactDetails can
belong to either a Person or an Organisation. To represent this within
the database schema it is easier to have the foreign keys on Person and
Organisation rather than on ContactDetails, even though this gives rise
to the inverse vocabulary within Rails - i.e, in the model file, a
Person "belongs_to" a ContactDetails.

Also, people will probably be looked up by name, so it is indexed, but
could share names, so it is not unique. They cannot share ContactDetails
so this can be uniquely indexed.