Add relationship between Organisation and People
To keep the API and data model simple, People can only belong to a
single Organisation. However, the relationship between Organisations
and People is its own table so that the relationship can be easily
extended, eg with multiple Organistions per Person or with metadata
about the relationship (title, date joined, etc).