Add Organisation APIs
This is heavily based on the Person APIs. Some common features have been
refactored into common superclasses or modules, but not all. The parts
that have been refactored out are those most likely to stay the same,
while those that have been copied are those that are most likely to
diverge in behaviour between the different model types. For example,
the common features of the API, such as error features, and the
interface to Contactables is likely to stay common between People and
Organisations, but the restrictions on how they are searched for, what
can be updated and deleted are likely not to remain the same.